+++ tests/AbstractProviderTest.php
@@ -2,8 +2,9 @@
 
 namespace Tests;
 
-use Ultraleet\CurrencyRates\Providers\DummyProvider;
+use Tests\Fixtures\TestProviderStub;
 use PHPUnit\Framework\TestCase as PHPUnit_Framework_TestCase;
+use DateTime;
 
 class AbstractProviderTest extends PHPUnit_Framework_TestCase
 {
@@ -11,7 +12,7 @@ class AbstractProviderTest extends PHPUnit_Framework_TestCase
 
     protected function setUp()
     {
-        $this->driver = new DummyProvider;
+        $this->driver = new TestProviderStub;
     }
 
     protected function tearDown()
@@ -24,5 +25,45 @@ class AbstractProviderTest extends PHPUnit_Framework_TestCase
         $driver = $this->driver->configure([]);
 
         $this->assertSame($this->driver, $driver);
+    }
+
+    public function testGetReturnsLatestResults()
+    {
+        $resultGet = $this->driver->get();
+        $resultLatest = $this->driver->latest();
+
+        $this->assertEquals($resultGet, $resultLatest);
+    }
+
+    public function testFluentDate()
+    {
+        $resultLatest = $this->driver->latest();
+        $resultHistorical = $this->driver->historical(new DateTime('2001-01-03'));
+
+        $resultDateString = $this->driver->date('2001-01-03')->get();
+        $resultDateDT = $this->driver->date(new DateTime('2001-01-03'))->get();
+        $resultDateNull = $this->driver->date()->get();
+
+        $this->assertEquals($resultHistorical, $resultDateString);
+        $this->assertEquals($resultHistorical, $resultDateDT);
+        $this->assertEquals($resultLatest, $resultDateNull);
+    }
+
+    public function testFluentBaseAndTarget()
+    {
+        $resultLatestBase = $this->driver->latest('EEK');
+        $resultLatestTarget = $this->driver->latest('EUR', ['EEK']);
+        $resultLatestTargets = $this->driver->latest('EUR', ['EEK', 'USD']);
+        $resultLatestBoth = $this->driver->latest('EEK', ['EUR']);
+
+        $resultBase = $this->driver->base('EEK')->get();
+        $resultTarget = $this->driver->base('EUR')->target('EEK')->get();           // string target
+        $resultTargets = $this->driver->base('EUR')->target(['EEK', 'USD'])->get(); // targets array
+        $resultBoth = $this->driver->base('EEK')->target(['EUR'])->get();
+
+        $this->assertEquals($resultLatestBase, $resultBase);
+        $this->assertEquals($resultLatestTarget, $resultTarget);
+        $this->assertEquals($resultLatestTargets, $resultTargets);
+        $this->assertEquals($resultLatestBoth, $resultBoth);
     }
 }